Выполнение задания
