+++ js/change-screen.js
@@ -1,23 +1,23 @@
 import {changeLevel} from './game.js';
 import {renderScreen} from './utils.js';
-import getGameScreen from './templates/game-screen.js';
-// import successScreen from './templates/success-screen.js';
-import resultScreen from './templates/result-screen.js';
-import failTriesScreen from './templates/fail-tries-screen.js';
+import resultScreen from './screens/result-screen.js';
+import failTriesScreen from './screens/fail-tries-screen.js';
+import gameScreen from './screens/game-screen.js';
 
-// Порядок смены экранов
+// Инкремент кол-ва жизней
+const LIVE_ADD = 1;
+
+// Смена экранов
 const changeScreen = (state) => {
-
-  // Если ещё есть жизни, то рендерим следующий экран, иначе переходим к определению результатов
   if (state.lives) {
     if (state.level < 9) {
-      const newScreen = changeLevel(state, state.level + 1);
-      renderScreen(getGameScreen(newScreen));
+      const newScreen = changeLevel(state, state.level + LIVE_ADD);
+      renderScreen(gameScreen(newScreen).element);
     } else {
-      renderScreen(resultScreen(state));
+      renderScreen(resultScreen(state).element);
     }
   } else {
-    renderScreen(failTriesScreen(state));
+    renderScreen(failTriesScreen(state).element);
   }
 
 };
