+++ js/main.js
@@ -1,5 +1,5 @@
 import {welcomeScreen} from './screens/welcome-screen.js';
-import {renderScreen} from './utils.js'; // добавляет элемент внутрь main.js
+import {renderScreen} from './utils.js';
 import {INITIAL_STATE} from './data';
 
 // Application.showWelcome();
