+++ js/screens/fail-tries-screen.js
@@ -1,6 +1,7 @@
 import {INITIAL_STATE} from '../data';
 import FailTriesView from '../views/fail-tries-view.js';
 import {renderScreen} from '../utils.js';
+import {welcomeScreen} from './welcome-screen.js';
 
 const failTriesScreen = (state) => {
   const screen = new FailTriesView(state);
