+++ js/screens/game-screen.js
@@ -1,5 +1,4 @@
 import GameView from '../views/game-view.js';
-import {renderElement} from '../utils.js';
 import header from '../screens/header.js';
 import genreScreen from '../screens/genre-screen.js';
 import artistScreen from '../screens/artist-screen.js';
@@ -14,7 +13,7 @@ const gameScreen = (state) => {
   // Контент экрана в зависимости от типа игры
   const content = (currentLevel.type === `game--artist`) ? artistScreen(state).element : genreScreen(state).element;
 
-  screen.element.insertAdjacentElement(`afterbegin`, gameHeader(state).element);
+  screen.element.insertAdjacentElement(`afterbegin`, header(state).element);
   screen.element.querySelector(`.game__screen`).appendChild(content);
 
   return screen;
